backend/services/recommendation-service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pandas
numpy>=1.24.2
scipy>=1.10.1
scikit-learn>=1.2.2
implicit>=0.6.1
azure-storage-blob>=12.16.0
azure-cosmos
python-dotenv
